# Add utility function to benchmark performance of fusion region with nvfuser and torch.compile