@kiya00 2025-01-22
>cc: @kiya00 I think this could be useful for automated report generation, would be great to have your review on this!

I think that's very great idea! I've prepared an initial version of the report, and we can gradually enrich its content https://github.com/Lightning-AI/lightning-thunder/pull/1636