@mruberry 2025-01-24
<img src="https://media2.giphy.com/media/XDicELf45BdEmmQ916/giphy.gif"/>

This looks ready to take out of draft to me, @kshitij12345, but maybe you have some more ideas

Running this and translating the slower fusions into bugs was incredibly interesting and helpful

fyi @kiya00, @riccardofelluga, I expect we'll want to develop the ability to have torch.compile or thunderfx run different fxgraphs, and maybe also the ability to autobenchmark torch.compile vs nvfuser on each fusion to select the faster of the two. We can talk about this more as we develop the new reporting tools